All path renames and symlinks
