+++ ruby/lib/cck/examples.rb
