+++ ruby/spec/cck/examples_spec.rb
